change Results into Formatters::Console
In the pursuit of eventually having other types of output
other than simply console output, I have taken the liberty of moving
the `Results` class under a module called `Formatters`.
